Move Base controller/views up
